+++ examples/display-blocking/src/main.rs
@@ -7,36 +7,16 @@ use panic_halt as _;
 use cortex_m_rt::entry;
 
 use microbit::{
-    display_pins,
-    hal::{gpio::p0::Parts as P0Parts, prelude::*, Timer},
+    board::Board,
+    display::blocking::Display,
+    hal::{prelude::*, Timer},
 };
-
-#[cfg(feature = "v2")]
-use microbit::hal::gpio::p1::Parts as P1Parts;
-
-use microbit::display::blocking::Display;
 
 #[entry]
 fn main() -> ! {
-    if let Some(p) = microbit::pac::Peripherals::take() {
-        let mut timer = Timer::new(p.TIMER0);
-
-        // Set up pins
-        #[cfg(feature = "v1")]
-        let pins = {
-            let p0parts = P0Parts::new(p.GPIO);
-            display_pins!(p0parts)
-        };
-
-        #[cfg(feature = "v2")]
-        let pins = {
-            let p0parts = P0Parts::new(p.P0);
-            let p1parts = P1Parts::new(p.P1);
-            display_pins!(p0parts, p1parts)
-        };
-
-        // Display
-        let mut display = Display::new(pins);
+    if let Some(board) = Board::take() {
+        let mut timer = Timer::new(board.TIMER0);
+        let mut display = Display::new(board.display_pins);
 
         #[allow(non_snake_case)]
         let letter_I = [