+++ microbit-common/src/board.rs
@@ -1,0 +1,6 @@
+//! Main Board
+#[cfg(feature = "v1")]
+pub use crate::v1::board::*;
+
+#[cfg(feature = "v2")]
+pub use crate::v2::board::*;